final commit

+++ LoaningApplication/Models/tblLoanDisbursementModel.cs
@@ -10,7 +10,7 @@ namespace LoaningApplication.Models
         public int disbursementID { get; set; }
         public int loanID { get; set; }
         public DateTime disbursementDate { get; set; }
-        public decimal disbursedAmount { get; set; }
+        public double disbursedAmount { get; set; }
         public string disbursementMethod { get; set; }
         public DateTime createAt { get; set; }
         public DateTime? updateAt { get; set; }
